fix: modify keybindings to account for Dvorak differences
Ghostty made a breaking change in May where some of their keybindings
are now defined using the PHYSICAL key on the keyboard versus the
logical key of one's input layout (Dvorak for example). So we have to
unbind those and rebind them using the single characeter instead of the
name.

See this discussion:
https://github.com/ghostty-org/ghostty/discussions/7540#discussioncomment-13402232