fix: Ghostty default bindings broke in dvorak
Ghostty made a breaking change in May 2025 where bindings defined with a
NAME (such as "equal") as opposed to a CHARACTER (such as "=") map to
the PHYSICAL key on the keyboard regardless of the input layout (dvorak
for example). So we need to unbind the mappings with names and replace
with character mappings if we want mappings that are consistent with
dvorak.